// BettingPlatform.WebApp/src/app/api/model/matchOfferDto.ts
/**
 * BettingPlatform.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MarketOfferDto } from './marketOfferDto';


export interface MatchOfferDto { 
    matchId?: string;
    homeTeam?: string | null;
    awayTeam?: string | null;
    startsAtUtc?: string;
    markets?: Array<MarketOfferDto> | null;
}