// BettingPlatform.WebApp/src/app/api/model/marketOfferDto.ts
/**
 * BettingPlatform.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OfferOutcomeDto } from './offerOutcomeDto';
import { OfferCategory } from './offerCategory';


export interface MarketOfferDto { 
    offerId?: string;
    marketCode?: string | null;
    category?: OfferCategory;
    outcomes?: Array<OfferOutcomeDto> | null;
}