// BettingPlatform.WebApp/src/app/api/model/offerDto.ts
/**
 * BettingPlatform.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OfferOutcomeDto } from './offerOutcomeDto';
import { OfferCategory } from './offerCategory';


export interface OfferDto { 
    offerId?: string;
    matchId?: string;
    homeTeam?: string | null;
    awayTeam?: string | null;
    startsAtUtc?: string;
    marketCode?: string | null;
    category?: OfferCategory;
    outcomes?: Array<OfferOutcomeDto> | null;
}